% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amprod.R
\name{amprod}
\alias{amprod}
\title{Amazon product data}
\usage{
amprod(number, version = 14, envir = parent.frame())
}
\arguments{
\item{number}{it refers to the number of the dataset you would like to use \code{\link{amprodlist}()}}
}
\description{
Amazon product data
}
\examples{
str(amprodlist)

}

