% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amprod.R
\docType{data}
\name{amprodlist}
\alias{amprodlist}
\title{Amazon product data}
\format{An object of class \code{list} of length 24.}
\usage{
amprodlist
}
\description{
Amazon product data
}
\details{
List of datasets per category
}
\examples{
1 = "Books"
2 = "Electronics"
3 = "Movies and TV"
4 = "CDs and Vinyl"
5 = "Clothing Shoes and Jewelry"
6 = "Home and Kitchen"
7 = "Kindle Store"
8 = "Sports and Outdoors"
9 = "Cell Phones and Accessories"
10 = "Health and Personal Care"
11 = "Toys and Games"
12 = "Video Games"
13 = "Tools and Home Improvement"
14 = "Beauty"
15 = "Apps for Android"
16 = "Office Products"
17 = "Pet Supplies"
18 = "Automotive"
19 = "Grocery and Gourmet Food"
20 = "Patio Lawn and Garden"
21 = "Baby"
22 = "Digital Music"
23 = "Musical Instruments"
24 = "Amazon Instant Video"


}
\keyword{datasets}

